closes #84, #86, #87 (simplify SPTextField, add support for CADisplayLink, add supporting rendering class)

+++ sparrow/src/Classes/SPStage.h
@@ -25,7 +25,7 @@
     id mNativeView;
 }
 
-@property (nonatomic, assign)   double frameRate;
+@property (nonatomic, assign)   float frameRate;
 @property (nonatomic, readonly) SPJuggler *juggler;
 @property (nonatomic, readonly) id nativeView;
